requirements.txt
Django==1.8.3
dj-database-url==0.3.0
dj-static==0.0.6
django-sslify==0.2.7
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
psycopg2==2.6.1
pystache==0.5.4
static3==0.6.1
